+++ index.php
@@ -365,8 +365,6 @@ elseif ( isset( $_REQUEST['create_account'] ) )
 
 		if ( ! UserSchool() )
 		{
-			RedirectURL( 'school_id' );
-
 			// @since 6.3 Create Student Account Default School.
 			// @link https://stackoverflow.com/questions/1250156/how-do-i-return-rows-with-a-specific-value-first#comment-67097263
 			$sql_order_by = Config( 'CREATE_STUDENT_ACCOUNT_DEFAULT_SCHOOL' ) ?