+++ CHANGES.md
@@ -65,6 +65,7 @@ Changes in 12.0
 - Add Greek & Turkish translations (99% completed) in locale/el_GR.utf8, locale/tr_TR.utf8 & REFERENCE.md, thanks to Georgios Katakalos
 - Deprecate isPopup() function & remove use in Warehouse() function in Warehouse.php
 - Fix do not use ProgramTitle() (loads Menu.php in English only) in Help.php
+- Update instructions, add pdo extension, remove deprecated config options in INSTALL.md, INSTALL_es.md & INSTALL_fr.md
 
 Changes in 11.8.6
 -----------------