+++ CHANGES.md
@@ -68,6 +68,7 @@ Changes in 12.0
 - Update instructions, add pdo extension, remove deprecated config options in INSTALL.md, INSTALL_es.md & INSTALL_fr.md
 - Deprecate Portal Notes Files upload path $PortalNotesFilesPath global var in PortalPollsNotes.fnc.php
 - Use $FileUploadsPath . 'PortalNotes/' instead of $PortalNotesFilesPath in PortalNotes.php
+- Remove Portal Notes Attached Files Folder in assets/PortalNotesFiles/
 
 Changes in 11.8.6
 -----------------