+++ CHANGES.md
@@ -71,6 +71,7 @@ Changes in 12.0
 - Remove Portal Notes Attached Files Folder in assets/PortalNotesFiles/
 - Deprecate Assignments Files upload path $AssignmentsFilesPath global var in StudentAssignments.fnc.php
 - Use $FileUploadsPath . 'Assignments/' instead of $AssignmentsFilesPath in StudentAssignments.fnc.php
+- Remove Student Assignments Files Folder in assets/AssignmentsFiles/ & .gitignore
 
 Changes in 11.8.6
 -----------------